Fix files structure to support go mod
