+++ packet.go
@@ -4,187 +4,171 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
+	"strings"
+)
+
+type Packet interface {
+	Unpack(data []byte) error
+	Pack() []byte
+}
+
+const (
+	OP_RRQ   = uint16(1) // Read request (RRQ)
+	OP_WRQ   = uint16(2) // Write request (WRQ)
+	OP_DATA  = uint16(3) // Data
+	OP_ACK   = uint16(4) // Acknowledgement
+	OP_ERROR = uint16(5) // Error
 )
 
 const (
-	opRRQ   = uint16(1) // Read request (RRQ)
-	opWRQ   = uint16(2) // Write request (WRQ)
-	opDATA  = uint16(3) // Data
-	opACK   = uint16(4) // Acknowledgement
-	opERROR = uint16(5) // Error
-	opOACK  = uint16(6) // Options Acknowledgment
+	BLOCK_SIZE        = 512
+	MAX_DATAGRAM_SIZE = 516
 )
 
-const (
-	blockLength    = 512
-	datagramLength = 516
-)
+type RRQ struct {
+	Filename string
+	Mode     string
+}
 
-type options map[string]string
-
-// RRQ/WRQ packet
-//
-//  2 bytes     string    1 byte    string    1 byte
-// --------------------------------------------------
-// | Opcode |  Filename  |   0  |    Mode    |   0  |
-// --------------------------------------------------
-type pRRQ []byte
-type pWRQ []byte
-
-// packRQ returns length of the packet in b
-func packRQ(p []byte, op uint16, filename, mode string, opts options) int {
-	binary.BigEndian.PutUint16(p, op)
-	n := 2
-	n += copy(p[2:len(p)-10], filename)
-	p[n] = 0
-	n++
-	n += copy(p[n:], mode)
-	p[n] = 0
-	n++
-	for name, value := range opts {
-		n += copy(p[n:], name)
-		p[n] = 0
-		n++
-		n += copy(p[n:], value)
-		p[n] = 0
-		n++
+func (p *RRQ) Unpack(data []byte) (e error) {
+	p.Filename, p.Mode, e = unpackRQ(data)
+	if e != nil {
+		return e
 	}
-	return n
+	return nil
 }
 
-func unpackRQ(p []byte) (filename, mode string, opts options, err error) {
-	bs := bytes.Split(p[2:], []byte{0})
-	if len(bs) < 2 {
-		return "", "", nil, fmt.Errorf("missing filename or mode")
+func (p *RRQ) Pack() []byte {
+	return packRQ(p.Filename, p.Mode, OP_RRQ)
+}
+
+type WRQ struct {
+	Filename string
+	Mode     string
+}
+
+func (p *WRQ) Unpack(data []byte) (e error) {
+	p.Filename, p.Mode, e = unpackRQ(data)
+	if e != nil {
+		return e
 	}
-	filename = string(bs[0])
-	mode = string(bs[1])
-	if len(bs) < 4 {
-		return filename, mode, nil, nil
+	return nil
+}
+
+func (p *WRQ) Pack() []byte {
+	return packRQ(p.Filename, p.Mode, OP_WRQ)
+}
+
+func unpackRQ(data []byte) (filename string, mode string, e error) {
+	buffer := bytes.NewBuffer(data[2:])
+	s, e := buffer.ReadString(0x0)
+	if e != nil {
+		return s, "", e
 	}
-	opts = make(options)
-	for i := 2; i+1 < len(bs); i += 2 {
-		opts[string(bs[i])] = string(bs[i+1])
+	filename = strings.TrimSpace(strings.Trim(s, "\x00"))
+	s, e = buffer.ReadString(0x0)
+	if e != nil {
+		return filename, s, e
 	}
-	return filename, mode, opts, nil
+	mode = strings.TrimSpace(s)
+	return filename, mode, nil
 }
 
-// OACK packet
-//
-// +----------+---~~---+---+---~~---+---+---~~---+---+---~~---+---+
-// |  Opcode  |  opt1  | 0 | value1 | 0 |  optN  | 0 | valueN | 0 |
-// +----------+---~~---+---+---~~---+---+---~~---+---+---~~---+---+
-type pOACK []byte
+func packRQ(filename string, mode string, opcode uint16) []byte {
+	buffer := &bytes.Buffer{}
+	binary.Write(buffer, binary.BigEndian, opcode)
+	buffer.WriteString(filename)
+	buffer.WriteByte(0x0)
+	buffer.WriteString(mode)
+	buffer.WriteByte(0x0)
+	return buffer.Bytes()
+}
 
-func packOACK(p []byte, opts options) int {
-	binary.BigEndian.PutUint16(p, opOACK)
-	n := 2
-	for name, value := range opts {
-		n += copy(p[n:], name)
-		p[n] = 0
-		n++
-		n += copy(p[n:], value)
-		p[n] = 0
-		n++
+type DATA struct {
+	BlockNumber uint16
+	Data        []byte
+}
+
+func (p *DATA) Unpack(data []byte) (e error) {
+	if len(data) < 4 { // DATA packet must have Opcode (2 bytes) and Block # (2 bytes)
+		return fmt.Errorf("invalid DATA packet (length = %d)", len(data))
 	}
-	return n
+	p.BlockNumber = binary.BigEndian.Uint16(data[2:])
+	p.Data = data[4:]
+	return nil
 }
 
-func unpackOACK(p []byte) (opts options, err error) {
-	bs := bytes.Split(p[2:], []byte{0})
-	opts = make(options)
-	for i := 0; i+1 < len(bs); i += 2 {
-		opts[string(bs[i])] = string(bs[i+1])
+func (p *DATA) Pack() []byte {
+	buffer := &bytes.Buffer{}
+	binary.Write(buffer, binary.BigEndian, OP_DATA)
+	binary.Write(buffer, binary.BigEndian, p.BlockNumber)
+	buffer.Write(p.Data)
+	return buffer.Bytes()
+}
+
+type ACK struct {
+	BlockNumber uint16
+}
+
+func (p *ACK) Unpack(data []byte) (e error) {
+	if len(data) < 4 { // ACK packet must have Opcode (2 bytes) and Block # (2 bytes)
+		return fmt.Errorf("invalid ACK packet (length = %d)", len(data))
 	}
-	return opts, nil
+	p.BlockNumber = binary.BigEndian.Uint16(data[2:])
+	return nil
 }
 
-// ERROR packet
-//
-//  2 bytes     2 bytes      string    1 byte
-// ------------------------------------------
-// | Opcode |  ErrorCode |   ErrMsg   |  0  |
-// ------------------------------------------
-type pERROR []byte
-
-func packERROR(p []byte, code uint16, message string) int {
-	binary.BigEndian.PutUint16(p, opERROR)
-	binary.BigEndian.PutUint16(p[2:], code)
-	n := copy(p[4:len(p)-2], message)
-	p[4+n] = 0
-	return n + 5
+func (p *ACK) Pack() []byte {
+	buffer := &bytes.Buffer{}
+	binary.Write(buffer, binary.BigEndian, OP_ACK)
+	binary.Write(buffer, binary.BigEndian, p.BlockNumber)
+	return buffer.Bytes()
 }
 
-func (p pERROR) code() uint16 {
-	return binary.BigEndian.Uint16(p[2:])
+type ERROR struct {
+	ErrorCode    uint16
+	ErrorMessage string
 }
 
-func (p pERROR) message() string {
-	return string(p[4:])
-}
-
-// DATA packet
-//
-//  2 bytes    2 bytes     n bytes
-// ----------------------------------
-// | Opcode |   Block #  |   Data   |
-// ----------------------------------
-type pDATA []byte
-
-func (p pDATA) block() uint16 {
-	return binary.BigEndian.Uint16(p[2:])
-}
-
-// ACK packet
-//
-//  2 bytes    2 bytes
-// -----------------------
-// | Opcode |   Block #  |
-// -----------------------
-type pACK []byte
-
-func (p pACK) block() uint16 {
-	return binary.BigEndian.Uint16(p[2:])
-}
-
-func parsePacket(p []byte) (interface{}, error) {
-	l := len(p)
-	if l < 2 {
-		return nil, fmt.Errorf("short packet")
+func (p *ERROR) Unpack(data []byte) (e error) {
+	p.ErrorCode = binary.BigEndian.Uint16(data[2:])
+	if len(data) < 4 { // ACK packet must have Opcode (2 bytes) and ErrorCode (2 bytes)
+		return fmt.Errorf("invalid ERROR packet (length = %d)", len(data))
 	}
-	opcode := binary.BigEndian.Uint16(p)
+	buffer := bytes.NewBuffer(data[4:])
+	s, e := buffer.ReadString(0x0)
+	if e != nil {
+		return e
+	}
+	p.ErrorMessage = strings.TrimSpace(s)
+	return nil
+}
+
+func (p *ERROR) Pack() []byte {
+	buffer := &bytes.Buffer{}
+	binary.Write(buffer, binary.BigEndian, OP_ERROR)
+	binary.Write(buffer, binary.BigEndian, p.ErrorCode)
+	buffer.WriteString(p.ErrorMessage)
+	buffer.WriteByte(0x0)
+	return buffer.Bytes()
+}
+
+func ParsePacket(data []byte) (Packet, error) {
+	var p Packet
+	opcode := binary.BigEndian.Uint16(data)
 	switch opcode {
-	case opRRQ:
-		if l < 4 {
-			return nil, fmt.Errorf("short RRQ packet: %d", l)
-		}
-		return pRRQ(p), nil
-	case opWRQ:
-		if l < 4 {
-			return nil, fmt.Errorf("short WRQ packet: %d", l)
-		}
-		return pWRQ(p), nil
-	case opDATA:
-		if l < 4 {
-			return nil, fmt.Errorf("short DATA packet: %d", l)
-		}
-		return pDATA(p), nil
-	case opACK:
-		if l < 4 {
-			return nil, fmt.Errorf("short ACK packet: %d", l)
-		}
-		return pACK(p), nil
-	case opERROR:
-		if l < 5 {
-			return nil, fmt.Errorf("short ERROR packet: %d", l)
-		}
-		return pERROR(p), nil
-	case opOACK:
-		if l < 6 {
-			return nil, fmt.Errorf("short OACK packet: %d", l)
-		}
-		return pOACK(p), nil
+	case OP_RRQ:
+		p = &RRQ{}
+	case OP_WRQ:
+		p = &WRQ{}
+	case OP_DATA:
+		p = &DATA{}
+	case OP_ACK:
+		p = &ACK{}
+	case OP_ERROR:
+		p = &ERROR{}
 	default:
 		return nil, fmt.Errorf("unknown opcode: %d", opcode)
 	}
+	return p, p.Unpack(data)
 }
