+++ v2/connection.go
