+++ v2/sender_anticipate.go
